3.2.8 getExtraBlocksFilms перенёс в utils
